third_party/blink/web_tests/external/wpt/css/css-lists/parsing/list-style-type-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['list-style-type'] = "symbols(cyclic \\"string\\")" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['list-style-type'] = "symbols(cyclic \\"○\\" \\"●\\")" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['list-style-type'] = "symbols(fixed \\"1\\")" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['list-style-type'] = "symbols(symbolic \\"string\\")" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['list-style-type'] = "symbols(alphabetic \\"first\\" \\"second\\")" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['list-style-type'] = "symbols(numeric \\"first\\" \\"second\\")" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
